Убрал в Layout регистрацию и логин
