+++ entities/LoginEntity.ts
@@ -1,7 +1,7 @@
 class LoginEntity {
   email!: string;
+
   password!: string;
 }
 
 export default LoginEntity;
-
